+++ src/InvokeRawCommandCommand.cs
@@ -1,6 +1,4 @@
-using System.Diagnostics.CodeAnalysis;
 using System.Management.Automation;
-using System.Management.Automation.Language;
 using Sashimi.Internal;
 
 namespace Sashimi;
@@ -16,7 +14,7 @@ public class InvokeRawCommandCommand : RawCommandBase
 
     [Parameter(ParameterSetName = NormalParameterSet, Mandatory = true, Position = 0,
                HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "InvokeRawCommand.parameters.Command")]
-    [ArgumentCompleter(typeof(NativeCommandCompleter))]
+    [ArgumentCompleter(typeof(ExternalCommandCompleter))]
     public string? Command { get; set; }
 
     [Parameter(ParameterSetName = NormalParameterSet, ValueFromRemainingArguments = true, Position = 1,
@@ -28,8 +26,8 @@ public class InvokeRawCommandCommand : RawCommandBase
     public ScriptBlock? Script { get; set; }
 
     [Parameter(ValueFromPipeline = true,
-               HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "InvokeRawCommand.parameters.InputBytes")]
-    public byte[]? InputBytes { get; set; }
+               HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "InvokeRawCommand.parameters.Input")]
+    public object? Input { get; set; }
 
     [Parameter(HelpMessageBaseName = MessageBaseName, HelpMessageResourceId = "InvokeRawCommand.parameters.Output")]
     [Alias("o")]
@@ -52,20 +50,11 @@ public class InvokeRawCommandCommand : RawCommandBase
         {
             if (Script is not null)
             {
-                if (SniffScriptBlock(Script, out var commandAst))
-                {
-                    var appInfo = GetAppInfo(commandAst.GetCommandName());
-                    _engine = new RawExecutionEngine(this, appInfo.Path);
-                }
-                else
-                {
-                    ThrowTerminatingError(new(new NotSupportedException("The ScriptBlock contains multiple statements or pipeline elements. "
-                                                                        + "Currently only a single external command is supported: raw { cmd }."),
-                                              "NotSupported",
-                                              ErrorCategory.NotImplemented,
-                                              this));
-                    return;
-                }
+                string[] forwardKeys = [nameof(Encoding), nameof(Output), nameof(AsString)];
+                var forwardParams = MyInvocation.BoundParameters
+                                    .Where(kv => forwardKeys.Contains(kv.Key, StringComparer.InvariantCulture))
+                                    .ToDictionary();
+                _engine = new ScriptBlockExecutionEngine(this, Script, forwardParams);
             }
             else if (Command is not null)
             {
@@ -91,9 +80,25 @@ public class InvokeRawCommandCommand : RawCommandBase
 
     protected override void ProcessRecord()
     {
-        if (InputBytes is not null)
+        if (Input is null)
+            return;
+
+        if (Input is PSObject pso)
+            Input = pso.BaseObject;
+
+        if (Input is string str)
         {
-            _engine?.ProcessRecord(InputBytes);
+            _engine?.ProcessRecord(str);
+        }
+        else if (LanguagePrimitives.TryConvertTo<byte[]>(Input, out var bytes))
+        {
+            _engine?.ProcessRecord(bytes);
+        }
+        else
+        {
+            var msg = $"Input data must be either byte[] or string: {Input?.GetType().Name ?? "null"}";
+            ThrowTerminatingError(new(new InvalidDataException(msg), "InvalidDataType", ErrorCategory.InvalidType, Input));
+            return;
         }
     }
 
@@ -116,43 +121,6 @@ public class InvokeRawCommandCommand : RawCommandBase
                                                   this));
         }
     }
-
-    private static bool SniffScriptBlock(ScriptBlock scriptBlock, [MaybeNullWhen(false)] out CommandAst commandAst)
-    {
-        commandAst = null;
-        var ast = scriptBlock.Ast as ScriptBlockAst;
-        if (ast is null)
-            return false;
-
-        if (ast.BeginBlock?.Statements.Count > 0)
-            return false;
-
-        if (ast.ProcessBlock?.Statements.Count > 0)
-            return false;
-
-        if (ast.EndBlock.Statements.Count != 1)
-            return false;
-
-        var pipeAst = ast.EndBlock.Statements[0] as PipelineAst;
-        if (pipeAst is null)
-            return false;
-
-        if (pipeAst.PipelineElements.Count != 1)
-            return false;
-
-        commandAst = pipeAst.PipelineElements[0] as CommandAst;
-
-        return commandAst is not null;
-    }
-
-    private static IEnumerable<string> GetArguments(CommandAst commandAst)
-        => commandAst.CommandElements.Skip(1)
-                                     .Select(elem => elem switch
-                                      {
-                                          StringConstantExpressionAst str => str.Value,
-                                          ExpandableStringExpressionAst exp => exp.Value,
-                                          _ => elem.Extent.Text
-                                      });
 
     private ApplicationInfo GetAppInfo(string name)
             => InvokeCommand.GetCommand(name, CommandTypes.Application) as ApplicationInfo